requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.2
setuptools==70.2.0
